src/guard/gptadapter/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html

python-jose==3.3.0
pika==1.3.2
fastapi==0.110.0
fastjsonschema==2.19.0
pandas==1.4.2
sqlalchemy==2.0.29
sqlalchemy-jsonfield==1.0.2
sqlalchemy-utils==0.41.1
sqlmodel==0.0.16
requests==2.31.0
numpy==1.22.4
pydantic==2.5.3
pydantic_core==2.14.6
scikit_learn==1.1.1
xgboost==1.6.1
psycopg2-binary==2.9.3
category-encoders==2.5.0
loguru==0.6.0
python-decouple==3.8
uvicorn[standard]
pytest~=8.1.1
starlette~=0.36.3
orjson~=3.9.15
passlib==1.7.4
python-multipart==0.0.9
httpx==0.26.0
watchdog==4.0.0
streamlit-cookies-controller
pytest==8.1.1
bcrypt
prometheus-client
prometheus-fastapi-instrumentator
yandexgpt-python
